+++ internal/appfeatures/update_lock.go
@@ -1,0 +1,5 @@
+package appfeatures
+
+type hostUpdateLock interface {
+	release() error
+}